Fixed argparse4j test errors
The main application used ArgumentParser.parseArgsOrFail(), which called
System.exit() on failure.  I replaced that call with .parseArgs() which
throws an ArgumentParserException instead.  That eliminates the need
for the SystemExitInterceptor which was trying to establish a (now
deprecated for removal) SecurityManager object.  This required that a
few tests were changed to catch a different exception.