Fix 'unused return value' warning
I think there might be more here that needs work.  It seems that the
try-catch block is inside an IF statement that checks whether
`savepath_dir.exists()` and `savepath_dir.isDirectory`.  The problem is
that it could exist and NOT be a directory -- I think? -- meaning that
the try-catch block will be skipped, so any error message won't happen
until much later, when the user tries to save the character...?  I don't
know the codebase well enough to know for sure.